+++ src/components/util/index.js
@@ -4,3 +4,4 @@ export * from './util';
 export { default as getRenewalPatronBlocksFromPatronBlocks } from './patronBlocks';
 export { nav };
 export { default as isRefundAllowed } from './isRefundAllowed';
+export { default as getLoanLastReminderNumber } from './getLoanLastReminderNumber';